Tests/LibWeb/Text/expected/async-scrolling/blocking-wheel-listener-invalidation.txt

before routing admission: accepted
listener generation increased: true
after routing admission: blocking wheel event listeners
